pyproject.toml
[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

[project]
name            = "chalkline"
version         = "0.1.0"
description     = "Career mapping for Maine's construction industry."
authors         = [{ name = "James Parkington" }]
readme          = "README.md"
requires-python = ">=3.14"
license         = { text = "MIT" }
dependencies    = [
    "htpy",
    "loguru",
    "marimo",
    "networkx",
    "pdfplumber",
    "plotly",
    "python-jobspy",
    "python-slugify",
    "onnxruntime",
    "scikit-learn",
    "sf-hamilton[diskcache]",
    "tokenizers",
    "typer"
]

[tool.uv]
override-dependencies = [
    "numpy>=2.0",
    "pandas>=3.0",
    "regex>=2025",
]

[project.scripts]
chalkline = "chalkline.cli:app"

[dependency-groups]
dev = [
    "nltk",
    "openpyxl",
    "pytest",
    "pytest-loguru",
    "wordfreq",
]
viz = [
    "seaborn",
]

[tool.hatch.build.targets.wheel]
packages = ["src/chalkline"]

[tool.marimo.runtime]
pythonpath = ["./src"]

[tool.pyright]
extraPaths = ["src"]

[tool.pytest.ini_options]
addopts    = "--import-mode=importlib"
cache_dir  = ".cache/pytest"
pythonpath = ["src"]
testpaths  = ["tests"]
